notifications/pyproject.toml
[project]
name = "notifications"
version = "0.1.0"
description = "Notifications service"
requires-python = ">=3.12"
dependencies = [
    "aiosmtplib>=3.0.2",
    "alembic>=1.13.0",
    "dishka>=1.3.0",
    "fastapi>=0.115.0",
    "faststream[kafka]>=0.5.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "pyjwt[crypto]>=2.9.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "uvicorn[standard]>=0.32.0",
    "websockets>=13.0",
    "psycopg[binary,pool]>=3.3.3",
    "structlog>=25.0.0",
    "redis>=7.2.0",
    "prometheus-fastapi-instrumentator>=7.1.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.2",
    "ty>=0.0.18",
]

[tool.ruff]
line-length = 120

exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    ".venv",
    "**/migrations/**",
]

[tool.ruff.lint]
select = [
    "ALL"
]
ignore = [
    "D103",
    "D102",
    "D100",
    "D101",
    "D104",
    "D107",
    "D105",
    "ANN401",
    "ANN204",
    "ANN202",
    "ANN201",
    "ANN003",
    "ANN002",
    "N815",
    "TC001",
    "D203",
    "D212",
    "PLR2004",
    "TRY301",
    "RUF006",
    "PGH003",
    "RUF002",
    "D415",
    "BLE001",
    "RUF001",
]


[tool.ruff.lint.isort]
known-first-party = [
    "application",
    "controllers",
    "infrastructure",
    "config",
    "ioc",
    "main",
]

[tool.ty.environment]
python-version = "3.12"
root = ["src"]

[tool.ty.src]
exclude = ["**/migrations/**"]
